cleanup items: renamings, add a unit test for transaction type util, extract public method from signing service
