+++ apps/policy-engine/src/resource/open-policy-agent/rego/__test__/criteria/approval_test.rego
@@ -1,12 +1,12 @@
 package main
 
 test_approversRoles {
-	roles = approversRoles with input as request with data.entities as entities
+	roles = approversRoles with input as requestWithEip1559Transaction with data.entities as entities
 	roles == {"root", "member", "admin"}
 }
 
 test_approversGroups {
-	groups = approversGroups with input as request with data.entities as entities
+	groups = approversGroups with input as requestWithEip1559Transaction with data.entities as entities
 	groups == {"test-user-group-one-uid", "test-user-group-two-uid"}
 }
 
@@ -17,7 +17,7 @@ test_checkApprovalByUserId {
 		"approvalEntityType": "Narval::User",
 		"entityIds": ["test-bob-uid", "test-bar-uid", "test-approver-uid"],
 	}
-	res = checkApproval(requiredApproval) with input as request with data.entities as entities
+	res = checkApproval(requiredApproval) with input as requestWithEip1559Transaction with data.entities as entities
 	res == 1
 }
 
@@ -28,7 +28,7 @@ test_checkApprovalByUserId {
 		"approvalEntityType": "Narval::User",
 		"entityIds": ["test-bob-uid", "test-bar-uid", "test-approver-uid"],
 	}
-	res = checkApproval(requiredApproval) with input as request with data.entities as entities
+	res = checkApproval(requiredApproval) with input as requestWithEip1559Transaction with data.entities as entities
 	res == 0
 }
 
@@ -39,7 +39,7 @@ test_checkApprovalByUserGroup {
 		"approvalEntityType": "Narval::UserGroup",
 		"entityIds": ["test-user-group-one-uid"],
 	}
-	res = checkApproval(requiredApproval) with input as request with data.entities as entities
+	res = checkApproval(requiredApproval) with input as requestWithEip1559Transaction with data.entities as entities
 	res == 1
 }
 
@@ -51,7 +51,7 @@ test_checkApprovalByUserGroup {
 		"entityIds": ["test-user-group-one-uid"],
 	}
 
-	res = checkApproval(requiredApproval) with input as request with data.entities as entities
+	res = checkApproval(requiredApproval) with input as requestWithEip1559Transaction with data.entities as entities
 
 	res == 0
 }
@@ -63,7 +63,7 @@ test_checkApprovalByUserRole {
 		"approvalEntityType": "Narval::UserRole",
 		"entityIds": ["root", "admin"],
 	}
-	res = checkApproval(requiredApproval) with input as request with data.entities as entities
+	res = checkApproval(requiredApproval) with input as requestWithEip1559Transaction with data.entities as entities
 	res == 2
 }
 
@@ -74,7 +74,7 @@ test_checkApprovalByUserRole {
 		"approvalEntityType": "Narval::UserRole",
 		"entityIds": ["root", "admin"],
 	}
-	res = checkApproval(requiredApproval) with input as request with data.entities as entities
+	res = checkApproval(requiredApproval) with input as requestWithEip1559Transaction with data.entities as entities
 	res == 3
 }
 
@@ -127,7 +127,7 @@ test_checkApprovals {
 		"entityIds": ["test-bob-uid", "test-bar-uid", "test-approver-uid"],
 	}
 
-	res = checkApprovals([satisfied, missing]) with input as request with data.entities as entities
+	res = checkApprovals([satisfied, missing]) with input as requestWithEip1559Transaction with data.entities as entities
 
 	res == {
 		"approvalsSatisfied": [satisfied],
