+++ apps/policy-engine/src/resource/open-policy-agent/rego/__test__/criteria/rateLimit_test.rego
@@ -79,7 +79,7 @@ test_calculateCurrentRateByRollingPeriod {
 		}
 	}
 
-	res = calculateCurrentRate(conditions) with input as request with data.entities as entities
+	res = calculateCurrentRate(conditions) with input as requestWithEip1559Transaction with data.entities as entities
     res == 2
 }
 
@@ -96,7 +96,7 @@ test_calculateCurrentRateByRollingPeriod {
 		}
 	}
 
-	res = calculateCurrentRate(conditions) with input as request with data.entities as entities
+	res = calculateCurrentRate(conditions) with input as requestWithEip1559Transaction with data.entities as entities
     res == 3
 }
 
@@ -113,7 +113,7 @@ test_calculateCurrentRateByRollingPeriod {
 		}
 	}
 
-	res = calculateCurrentRate(conditions) with input as request with data.entities as entities
+	res = calculateCurrentRate(conditions) with input as requestWithEip1559Transaction with data.entities as entities
     res == 1
 }
