+++ apps/policy-engine/src/resource/open-policy-agent/rego/__test__/policies/approvals_test.rego
@@ -91,7 +91,7 @@ test_withoutApprovalsEIP1559 {
 	}
 }
 
-test_withoutApprovalsLEGACY {
+test_withoutApprovalsLegacy {
 	withoutApprovalsReq := {
 		"action": "signTransaction",
 		"transactionRequest": transactionRequestLegacy,
